requirements.txt
flask
python-dotenv
requests

gunicorn
redis

openpyxl
polars
watchdog

pandas==2.1.4
numpy==1.26.4

openai

llama-index
llama-index-core
llama-index-agent-openai
llama-index-llms-openai
llama-index-embeddings-openai
